requirements-dev.txt
# Include production requirements
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.5.0

# Code quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
pylint>=3.0.0
mypy>=1.7.0

# Type stubs
types-requests>=2.31.0
types-PyYAML>=6.0.12
pandas-stubs>=2.1.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
sphinx-autodoc-typehints>=1.25.0

# Development tools
ipython>=8.18.0
ipdb>=0.13.13
pre-commit>=3.6.0

# Build tools
setuptools>=69.0.0
wheel>=0.42.0
build>=1.0.0
twine>=4.0.0

# Security
bandit>=1.7.5
safety>=2.3.0

# Performance profiling
line-profiler>=4.1.0
memory-profiler>=0.61.0

# Linting
pycodestyle>=2.11.0
pydocstyle>=6.3.0
